+++ tests/project.rs
@@ -95,6 +95,8 @@ mod manifest_v10;
 mod manifest_v11;
 #[path = "project/manifest_v12.rs"]
 mod manifest_v12;
+#[path = "project/manifest_v13.rs"]
+mod manifest_v13;
 #[path = "project/manifest_v4.rs"]
 mod manifest_v4;
 #[path = "project/manifest_v5.rs"]